Refactor: Move components around.
