+++ src/components/Branding/index.js
@@ -2,4 +2,3 @@ export { default as Ava } from './Ava';
 export { default as Indicator } from './Indicator';
 export { default as Logo } from './Logo';
 export { default as Logotype } from './Logotype';
-export { default as Train } from './Train';
