+++ src/components/Header.js
@@ -3,10 +3,9 @@ import PropTypes from 'prop-types';
 
 import styled from 'styled-components';
 
-import Logo from './Branding/Logo';
-import Logotype from './Branding/Logotype';
-import Timer from './Branding/Timer';
-import Train from './Branding/Train';
+import { Logo, Logotype } from './Branding';
+import { Timer, Train } from './Train';
+import { Pool } from './Pool';
 
 function Hero({ className }) {
   return (
@@ -15,6 +14,7 @@ function Hero({ className }) {
       <StyledLogo />
       <StyledTrain />
       <StyledLogotype />
+      <StyledPool />
     </Wrapper>
   );
 }
@@ -29,13 +29,13 @@ Hero.defaultProps = {
 
 const Wrapper = styled.div`
   display: grid;
-  grid-template-columns: auto 1fr;
+  grid-template-columns: auto auto 1fr;
   grid-template-rows: 24px auto;
   align-items: center;
 `;
 
 const StyledTimer = styled(Timer)`
-  grid-column: 1 / span 2;
+  grid-column: 1 / span 3;
   grid-row: 1;
   align-self: start;
 `;
@@ -52,6 +52,11 @@ const StyledTrain = styled(Train)`
 
 const StyledLogotype = styled(Logotype)`
   grid-column: 2;
+  grid-row: 2;
+`;
+
+const StyledPool = styled(Pool)`
+  grid-column: 3;
   grid-row: 2;
 `;
