+++ src/components/Train/index.js
@@ -1,0 +1,2 @@
+export { default as Train } from './Train';
+export { default as Timer } from './Timer';
