+++ src/components/index.js
@@ -1,9 +1,10 @@
-export { Ava, Logo, Logotype, Train } from './Branding';
+export { Ava, Logo, Logotype } from './Branding';
 export { default as Camera } from './Camera';
 export { default as Debug } from './Debug';
 export { default as Header } from './Header';
 export { Notifier } from './Notifier';
 export { Queue } from './Queue';
 export { Pool } from './Pool';
-export { Ticker } from './Ticker';
 export { Summaries } from './Summary';
+export { Ticker } from './Ticker';
+export { Timer, Train } from './Train';
