+++ src/scenes/Activity/Activity.js
@@ -1,15 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {
-  Debug,
-  Header,
-  Notifier,
-  Pool,
-  Queue,
-  Summaries,
-  Ticker
-} from 'components';
+import { Debug, Header, Notifier, Queue, Summaries, Ticker } from 'components';
 import { useNotificationContext } from 'providers';
 
 import styled from 'styled-components';
@@ -20,7 +12,6 @@ function Layout() {
 
   return (
     <StyledWrapper>
-      <Pool />
       <StyledHeader notifications={notifications} />
       <StyledSummaries isVisible />
       <StyledTicker isVisible />
